+++ Editor/CustomEditors/BehaviourTreeInstanceEditor.cs
@@ -16,10 +16,18 @@ namespace TheKiwiCoder {
             PropertyField validateField = new PropertyField();
             validateField.bindingPath = nameof(BehaviourTreeInstance.validate);
 
+            PropertyField tickMode = new PropertyField();
+            tickMode.bindingPath = nameof(BehaviourTreeInstance.tickMode);
+
+            PropertyField startMode = new PropertyField();
+            startMode.bindingPath = nameof(BehaviourTreeInstance.startMode);
+
             PropertyField publicKeys = new PropertyField();
             publicKeys.bindingPath = nameof(BehaviourTreeInstance.blackboardOverrides);
 
             container.Add(treeField);
+            container.Add(tickMode);
+            container.Add(startMode);
             container.Add(validateField);
             container.Add(publicKeys);