+++ Runtime/Actions/RandomPosition.cs
@@ -8,10 +8,10 @@ namespace TheKiwiCoder {
     public class RandomPosition : ActionNode {
 
         [Tooltip("Minimum bounds to generate point")] 
-        public Vector2 min = Vector2.one * -10;
+        public Vector3 min = Vector2.one * -10;
 
         [Tooltip("Maximum bounds to generate point")] 
-        public Vector2 max = Vector2.one * 10;
+        public Vector3 max = Vector2.one * 10;
 
         [Tooltip("Blackboard key to write the result to")] 
         public NodeProperty<Vector3> result;
@@ -26,6 +26,7 @@ namespace TheKiwiCoder {
             Vector3 pos = new Vector3();
             pos.x = Random.Range(min.x, max.x);
             pos.y = Random.Range(min.y, max.y);
+            pos.z = Random.Range(min.z, max.z);
             result.Value = pos;
             return State.Success;
         }