+++ Runtime/Composites/Sequencer.cs
@@ -5,27 +5,21 @@ using UnityEngine;
 namespace TheKiwiCoder {
     [System.Serializable]
     public class Sequencer : CompositeNode {
-        protected int current;
 
         protected override void OnStart() {
-            current = 0;
         }
 
         protected override void OnStop() {
         }
 
         protected override State OnUpdate() {
-            for (int i = current; i < children.Count; ++i) {
-                current = i;
-                var child = children[current];
-
-                switch (child.Update()) {
-                    case State.Running:
-                        return State.Running;
-                    case State.Failure:
-                        return State.Failure;
-                    case State.Success:
-                        continue;
+            for (int i = 0; i < children.Count; ++i) {
+                var childStatus = children[i].Update();
+                
+                if (childStatus == State.Running) {
+                    return State.Running;
+                } else if (childStatus == State.Failure) {
+                    return State.Failure;
                 }
             }